Add pandas dir
