+++ packages/client/libclient-py/quickmpc/__init__.py
@@ -1,6 +1,6 @@
+from .pandas.share_data_frame import ShareDataFrame
 from .qmpc import QMPC
 from .qmpc_logging import get_logger  # noqa: F401
-from .share_data_frame import ShareDataFrame
 from .version import __version__
 
 __all__ = [
