+++ packages/client/libclient-py/quickmpc/pandas/parser.py
@@ -8,7 +8,7 @@ import numpy as np
 from ..exception import ArgumentError
 from ..proto.common_types.common_types_pb2 import Schema, ShareValueTypeEnum
 from ..qmpc_logging import get_logger
-from .overload_tools import Dim1, methoddispatch
+from ..utils.overload_tools import Dim1, methoddispatch
 
 logger = get_logger()
