+++ packages/client/libclient-py/quickmpc/qmpc_request.py
@@ -19,6 +19,7 @@ import tqdm  # type: ignore
 from grpc_status import rpc_status  # type: ignore
 
 from .exception import ArgumentError, QMPCJobError, QMPCServerError
+from .pandas.parser import parse
 from .proto.common_types.common_types_pb2 import (ComputationMethod,
                                                   JobErrorInfo, Schema)
 from .proto.libc_to_manage_pb2 import (AddShareDataFrameRequest,
@@ -37,7 +38,6 @@ from .request.response import (AddShareDataFrameResponse, ExecuteResponse,
 from .share import Share
 from .utils.if_present import if_present
 from .utils.make_pieces import MakePiece
-from .utils.parse_csv import parse
 
 logger = logging.getLogger(__name__)
