+++ packages/client/libclient-py/tests/unit_tests/test_qmpc.py
@@ -5,14 +5,9 @@ import pandas as pd
 import pytest
 
 from quickmpc.exception import ArgumentError
+from quickmpc.pandas.share_data_frame import ShareDataFrameStatus
 from quickmpc.qmpc import QMPC
 from quickmpc.qmpc_request import QMPCRequest
-from quickmpc.share_data_frame import ShareDataFrameStatus
-
-
-def to_string_io(data: List[List]) -> io.StringIO:
-    text_data = "\n".join([",".join(map(str, row)) for row in data])
-    return io.StringIO(text_data)
 
 
 def data_frame(values: List[List] = [[1, 2], [3, 4]],
@@ -61,33 +56,6 @@ class TestQMPC:
                               run_server1, run_server2, run_server3):
         with pytest.raises(expected):
             QMPC(arg)
-
-    @pytest.mark.parametrize(
-        ("data", "index_col", "expected"), [
-            # 通常パターン
-            ([["id", "c"], ["a", 1], ["b", 2]],
-             "id",
-             pd.DataFrame([[32772040.0, 1], [86407020.0, 2]],
-                          columns=["id", "c"],
-                          index=[0, 1])),
-            # 座圧されたID列がindexにあるかどうか
-            ([["id", "c"], ["b", 2], ["a", 1]],
-             "id",
-             pd.DataFrame([[86407020.0, 2], [32772040.0, 1]],
-                          columns=["id", "c"],
-                          index=[1, 0])),
-            # 1列目以外をID列に指定した場合
-            ([["id", "c"], ["a", 1], ["b", 2]],
-             "c",
-             pd.DataFrame([["a", 1.0], ["b", 2.0]],
-                          columns=["id", "c"],
-                          index=[0, 1])),
-        ]
-    )
-    def test_read_csv(self, data, index_col, expected,
-                      run_server1, run_server2, run_server3):
-        df = self.qmpc.read_csv(to_string_io(data), index_col=index_col)
-        pd.testing.assert_frame_equal(df, expected)
 
     def test_send_to(self,
                      run_server1, run_server2, run_server3):
