+++ packages/client/libclient-py/tests/unit_tests/test_share_data_frame.py
@@ -8,8 +8,9 @@ from typing import List, Optional
 import pandas as pd
 import pytest
 
+from quickmpc.pandas.share_data_frame import (ShareDataFrame,
+                                              ShareDataFrameStatus)
 from quickmpc.qmpc_request import QMPCRequest
-from quickmpc.share_data_frame import ShareDataFrame, ShareDataFrameStatus
 
 
 def to_string_io(data: List[List]) -> io.StringIO:
